slice 1 redux slice on 3 slices
