+++ src/store/index.ts
@@ -1,8 +1,14 @@
 import { configureStore } from '@reduxjs/toolkit'
+import burgerConstructorSlice from './burgerConstructorSlice'
 import ingridientsSlice from './ingridientsSlice'
+import orderSlice from './orderSlice'
 
 const store = configureStore({
-  reducer: { ingridients: ingridientsSlice },
+  reducer: {
+    ingridients: ingridientsSlice,
+    burgerConstructor: burgerConstructorSlice,
+    order: orderSlice,
+  },
 })
 
 export default store
