+++ src/types/burgerConstructorTypes.ts
@@ -1,0 +1,5 @@
+import { Ingridient } from './ingridient'
+
+export interface BurgerConstructorSlice {
+  inConstructor: Array<Ingridient>
+}
