+++ src/types/ingridientsTypes.ts
@@ -1,0 +1,12 @@
+import { Ingridient } from './ingridient'
+
+export interface IngridientsSlice {
+  loading: boolean
+  error: string | null | undefined
+  ingredients: Array<Ingridient>
+}
+
+export interface IngridientsData {
+  success: boolean
+  data: Ingridient[]
+}
